Add a controller with click buttons and simple operations
Project: measurecalcexample

Add a controller with click buttons and simple operations. Curently, the calculator only calculates tons to grams.

+++ measurecalcexample/src/sample/Calculator.java
@@ -10,7 +10,7 @@ public class Calculator {
         return instance;
     }
 
-    public void setNumberToCalculate(int numberToCalculate) {
+    public void setNumberToCalculate(double numberToCalculate) {
         this.numberToCalculate = numberToCalculate;
     }
